backend-python/requirements.txt
mysql-connector-python==8.1.0
pandas==2.1.0
azure-mgmt-network==27.0.0
azure-identity==1.15.0
openai==1.10.0
python-dotenv==1.0.0
protobuf>=5.28.0
